pyproject.toml
[project]
name = "pieces-auto-tn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.128.8",
    "httpx>=0.28.1",
    "playwright>=1.58.0",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.7",
    "uvicorn>=0.39.0",
]
